test_aoai/requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
python-dotenv
pydantic
openai
pypdf
rank-bm25
pandas